Add ThreaddedNetwork and make code work on Linux
Created a new class called threadded network that runs the networking
work in a seperate thread and leave the main thread free to run.  when
new messages come in they are placed in a queue that can be accessed in the
main thread.

added some includes to make the code run on thor